# perf(core/ops): avoid allocs when returning primitives