--- a/core/benches/op_baseline.rs
+++ b/core/benches/op_baseline.rs
@@ -4,9 +4,9 @@ use deno_core::error::AnyError;
 use deno_core::op_async;
 use deno_core::op_sync;
 use deno_core::v8;
+use deno_core::serialize_op_result;
 use deno_core::JsRuntime;
 use deno_core::Op;
-use deno_core::OpResponse;
 use deno_core::OpState;
 use deno_core::ZeroCopyBuf;
 
@@ -18,7 +18,7 @@ fn create_js_runtime() -> JsRuntime {
   runtime.register_op("pi_json", op_sync(|_, _: (), _| Ok(314159)));
   runtime.register_op("pi_async", op_async(op_pi_async));
   runtime
-    .register_op("nop", |_, _, _| Op::Sync(OpResponse::Value(Box::new(9))));
+    .register_op("nop", |state, _, _| Op::Sync(serialize_op_result(Ok(9), state)));
 
   // Init ops
   runtime