--- a/core/lib.rs
+++ b/core/lib.rs
@@ -5,6 +5,7 @@ mod bindings;
 pub mod error;
 mod flags;
 mod gotham_state;
+mod minvalue;
 mod module_specifier;
 mod modules;
 mod normalize_path;