--- a/core/ops.rs
+++ b/core/ops.rs
@@ -4,6 +4,7 @@ use crate::error::bad_resource_id;
 use crate::error::type_error;
 use crate::error::AnyError;
 use crate::gotham_state::GothamState;
+use crate::minvalue::SerializablePkg;
 use crate::resources::ResourceId;
 use crate::resources::ResourceTable;
 use crate::runtime::GetErrorClassFn;
@@ -61,7 +62,7 @@ impl<'a, 'b, 'c> OpPayload<'a, 'b, 'c> {
 }
 
 pub enum OpResponse {
-  Value(Box<dyn serde_v8::Serializable>),
+  Value(OpResult),
   Buffer(Box<[u8]>),
 }
 
@@ -74,11 +75,22 @@ pub enum Op {
   NotFound,
 }
 
-#[derive(Serialize)]
-#[serde(untagged)]
-pub enum OpResult<R> {
-  Ok(R),
+pub enum OpResult {
+  Ok(SerializablePkg),
   Err(OpError),
+}
+
+impl OpResult {
+  pub fn to_v8<'a>(
+    &self,
+    scope: &mut v8::HandleScope<'a>,
+  ) -> Result<v8::Local<'a, v8::Value>, serde_v8::Error> {
+    
+    match self {
+      Self::Ok(x) => x.to_v8(scope),
+      Self::Err(err) => serde_v8::to_v8(scope, err),
+    }
+  }
 }
 
 #[derive(Serialize)]
@@ -93,13 +105,13 @@ pub fn serialize_op_result<R: Serialize + 'static>(
   result: Result<R, AnyError>,
   state: Rc<RefCell<OpState>>,
 ) -> OpResponse {
-  OpResponse::Value(Box::new(match result {
-    Ok(v) => OpResult::Ok(v),
+  OpResponse::Value(match result {
+    Ok(v) => OpResult::Ok(v.into()),
     Err(err) => OpResult::Err(OpError {
       class_name: (state.borrow().get_error_class_fn)(&err),
       message: err.to_string(),
     }),
-  }))
+  })
 }
 
 /// Maintains the resources and ops inside a JS runtime.