--- a/core/ops.rs
+++ b/core/ops.rs
@@ -85,7 +85,6 @@ impl OpResult {
     &self,
     scope: &mut v8::HandleScope<'a>,
   ) -> Result<v8::Local<'a, v8::Value>, serde_v8::Error> {
-    
     match self {
       Self::Ok(x) => x.to_v8(scope),
       Self::Err(err) => serde_v8::to_v8(scope, err),