This introduces `SerializablePkg` and `MinValue` as a fastpath for ops that return primitives, allowing us to save an alloc/free per opcall. This reduces async-op baseline overhead by ~20% or ~100ns/opcall.

⚠️ **Warning:** This uses `unsafe` code (`std::mem::transmute_copy`) and TypeId checks in lieu of specialization to detect primitives and opt-in to the fast path

## Benches

```
Before:
test bench_op_async   ... bench:     489,671 ns/iter (+/- 8,368)
test bench_op_nop     ... bench:      71,506 ns/iter (+/- 1,612)
test bench_op_pi_json ... bench:      84,846 ns/iter (+/- 1,464)

After:
test bench_op_async   ... bench:     384,175 ns/iter (+/- 7,553)
test bench_op_nop     ... bench:      47,984 ns/iter (+/- 2,083)
test bench_op_pi_json ... bench:      59,100 ns/iter (+/- 1,868)
```